+++ CHANGES.md
@@ -11,6 +11,7 @@ Since last release
 #### 11/29/17
   - Fix typo preventing `pardiso.dparm` options from being set.
   - Fix incorrect use of PARDISO options in `t_mplinsolve()`.
+  - Add preliminary support for PARDISO v6.x.
 
 #### 3/16/17
   - Move `mplinsolve` PARDISO options to `opt.pardiso` in preparation